+++ demo/demo/settings.py
@@ -40,6 +40,7 @@ INSTALLED_APPS = [
 
     # Require
     'bootstrap3',
+    'bootstrap4',
     'crispy_forms',
 
     # Added extra