+++ django_popup_view_field/__init__.py
@@ -1,2 +1,3 @@
-VERSION = (0, 5, 0)
+VERSION = (0, 6, 0)
 __version__ = ".".join(str(i) for i in VERSION)
+SUPPORTED_BOOTSTRAP_VER = ('bootstrap3', 'bootstrap4')