+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG for django-popup-view-field
 
+## 0.6.0 (2020-01-23)
+
+* Added support for Django 2.2
+
+* Drop support for Django 1.8
+
+* Added support for bootstrap4 (testing only with Django >= 2.1)
+
+* Added setting DJANGO_POPUP_VIEW_FIELD_TEMPLATE_PACK
+
+* Templates for widgets was splitted to blocks
+  (for easy customization). Now `PopupViewWidget` can use
+  template for bootstrap3 or template for bootstrap4
+
 ## 0.5.0 (2019-04-17)
 
 * Support for Django 2.0 and 2.1